judger add linux timeout and js bug fixed

+++ judger.py
@@ -20,24 +20,23 @@ def rundocker(judging):
     to2 = "/tmp/code"
     if judging.problem.test=="":
         if judging.lang == "python3":
-            bash = "\"> tmp/code/error.txt;python3 tmp/code/* > tmp/code/out.txt 2>>tmp/code/error.txt\""
+            bash = "\"> tmp/code/error.txt;timeout 10 python3 tmp/code/*.py > tmp/code/out.txt 2>>tmp/code/error.txt;echo $?>tmp/code/timeout.txt\""
         elif judging.lang == "c++":
-            bash = "\"> tmp/code/error.txt;g++ tmp/code/*.cpp -o tmp/code/a.exe 2> tmp/code/error.txt;tmp/code/a.exe > tmp/code/out.txt 2>> tmp/code/error.txt\""
+            bash = "\"> tmp/code/error.txt;g++ tmp/code/*.cpp -o tmp/code/a.exe 2> tmp/code/error.txt;timeout 10 tmp/code/a.exe > tmp/code/out.txt 2>> tmp/code/error.txt;echo $?>tmp/code/timeout.txt\""
     else:
         if judging.lang == "python3":
-            bash = "\"> tmp/code/error.txt;python3 tmp/code/* < tmp/problem/test.txt > tmp/code/out.txt 2>>tmp/code/error.txt\""
+            bash = "\"> tmp/code/error.txt;timeout 10 python3 tmp/code/*.py < tmp/problem/test.txt > tmp/code/out.txt 2>>tmp/code/error.txt;echo $?>tmp/code/timeout.txt\""
         elif judging.lang == "c++":
-            bash = "\"> tmp/code/error.txt;g++ tmp/code/*.cpp -o tmp/code/a.exe 2> tmp/code/error.txt;tmp/code/a.exe < tmp/problem/test.txt > tmp/code/out.txt 2>>tmp/code/error.txt\""
+            bash = "\"> tmp/code/error.txt;g++ tmp/code/*.cpp -o tmp/code/a.exe 2> tmp/code/error.txt;timeout 10 tmp/code/a.exe < tmp/problem/test.txt > tmp/code/out.txt 2>>tmp/code/error.txt;echo $?>tmp/code/timeout.txt\""
     cmd = 'docker run -dit -v %s:%s:z -v %s:%s:z --privileged=true --memory="64M" --memory-swap="64M" --cpu-quota=75000 --name %s charlie890414/judger:latest bash -c %s' % (problem,to1,code,to2,name,bash)
     print(cmd)
-    tmp = subprocess.Popen(cmd)
-    for i in range(11):        
-        time.sleep(1)
-        if tmp.poll()!=None:
-            break
-        elif  i == 10:
-            tmp.terminate()
-            submission.objects.filter(id=judging.id).update(status='TLE')
+    tmp = subprocess.call(cmd)
+    time.sleep(3)
+    timeout =open(code+"\\timeout.txt").read()
+    print(timeout)
+    if timeout == '124':
+        submission.objects.filter(id=judging.id).update(status='TLE')
+        return
     cmd = 'docker stop %s' %(name)
     tmp = subprocess.call(cmd)
     cmd = 'docker rm %s' %(name)
